Single channel logging (#3)
* Moved global variables in FP_acquire into a struct

* Hibernate wake reconfiguration

Realized that when an RTC match occurs it starts running main().

* End of Day

Not great but at least one measurement is read in now

* End of Day Nov 23

Logging works if ADC parts are uncommented. Was trying to get speed up to 2.5kHz but I think UART is limiting it (not sure).

* Version for push to master

Logs Ch0 at up to 600 Hz with a subsecond timestamp.